+++ libs/cosmos-sdk/x/bank/internal/types/msgs.go
@@ -85,7 +85,7 @@ func (msg *MsgSend) UnmarshalFromAmino(cdc *amino.Codec, data []byte) error {
 			return err
 		}
 		data = data[n:]
-		if len(data) < int(dataLen) {
+		if uint64(len(data)) < dataLen {
 			return fmt.Errorf("invalid tx data")
 		}
 		subData = data[:dataLen]
